Extract SearchBox into a new component
